Remove wallet interface
