+++ src/main/java/eu/qwan/vender/ChipknipWallet.java
@@ -1,6 +1,6 @@
 package eu.qwan.vender;
 
-public class ChipknipWallet extends WalletImpl {
+public class ChipknipWallet extends Wallet {
 
     public ChipknipWallet(int initial_value) {
         credits = initial_value;
