+++ src/main/java/eu/qwan/vender/CoinWallet.java
@@ -1,4 +1,4 @@
 package eu.qwan.vender;
 
-public class CoinWallet extends WalletImpl {
+public class CoinWallet extends Wallet {
 }
